[maida Ge] feat: add vote api for vote a rs event

+++ src/main/java/com/thoughtworks/rslist/service/RsServiceImpl.java
@@ -1,5 +1,8 @@
 package com.thoughtworks.rslist.service;
 
+import com.thoughtworks.rslist.Exception.InvalidIndexException;
+import com.thoughtworks.rslist.Exception.InvalidRequestParamException;
+import com.thoughtworks.rslist.Exception.UserNotExistedException;
 import com.thoughtworks.rslist.domain.RsEvent;
 import com.thoughtworks.rslist.dto.RsEventDto;
 import com.thoughtworks.rslist.dto.UserDto;
@@ -10,8 +13,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
 @Service
@@ -31,7 +32,7 @@ public class RsServiceImpl implements RsService{
     public int addRsEvent(RsEvent rsEvent) {
         Optional<UserDto> userDto = userRepository.findById(rsEvent.getUserId());
         if (!userDto.isPresent()) {
-            return -1;
+            throw new InvalidIndexException();
         }
         RsEventDto rsEventDto = rsEventRepository.save(rsEvent.toRsEventDto(userDto.get()));
         return rsEventDto.getId();
@@ -45,14 +46,17 @@ public class RsServiceImpl implements RsService{
     @Override
     public RsEvent getRsEvent(int id) {
         Optional<RsEventDto> rsEventDto = rsEventRepository.findById(id);
-        if(rsEventDto.isPresent()) {
-            return rsEventDto.get().toRsEvent();
+        if(!rsEventDto.isPresent()) {
+            throw new InvalidIndexException();
         }
-        return null;
+        return rsEventDto.get().toRsEvent();
     }
 
     @Override
     public List<RsEvent> getRsEventBetween(int start, int end) {
+        if (start < 1 || end < 1 || end < start) {
+            throw new InvalidRequestParamException();
+        }
         return rsEventRepository.findAllByIdBetween(start, end)
                 .stream()
                 .map(rsEventDto -> rsEventDto.toRsEvent())
@@ -63,11 +67,12 @@ public class RsServiceImpl implements RsService{
     public int updateRsEvent(RsEvent rsEvent, int id) {
         Optional<RsEventDto> rsEventDto = rsEventRepository.findById(id);
         if (!rsEventDto.isPresent()) {
-            return -1;
+            throw new InvalidIndexException();
         }
         if (rsEventDto.get().getUserDto().getId() != rsEvent.getUserId()) {
-            return -1;
+            throw new UserNotExistedException();
         }
+
         String updateName = rsEvent.getEventName();
         String updateKeyWord = rsEvent.getKeyWord();
         int affectRow = 0;
@@ -82,18 +87,14 @@ public class RsServiceImpl implements RsService{
 
     @Override
     public void deleteRsEvent(int id) {
+        if (!rsEventRepository.findById(id).isPresent()) {
+            throw new InvalidIndexException();
+        }
         rsEventRepository.deleteById(id);
     }
 
     @Override
     public int getRsNumber() {
         return Integer.parseInt(String.valueOf(rsEventRepository.count()));
-    }
-
-    @Override
-    public int getMaxId() {
-        AtomicInteger maxId = new AtomicInteger(0);
-        rsEventRepository.findAll().forEach(rsEventDto -> maxId.set(rsEventDto.getId() > maxId.get() ? rsEventDto.getId() : maxId.get()));
-        return maxId.intValue();
     }
 }
